Backport pull request #2366 from erikn69/patch-7 Fix delete permissions on Permissions Model

+++ src/Traits/HasPermissions.php
@@ -32,6 +32,10 @@ trait HasPermissions
                 return;
             }
 
+            if (is_a($model, Permission::class)) {
+                return;
+            }
+
             $teams = PermissionRegistrar::$teams;
             PermissionRegistrar::$teams = false;
             $model->permissions()->detach();
